ENG-13317: The static modified counter in CRUD executors should only be initialized to -1 when executing a replicated table action on the lowest site thread. Otherwise the assignment can race with a slow non-lowest site thread that has not finished the last MP txn and thinks incorrectly that an exception needs to be thrown.

+++ src/ee/executors/swaptablesexecutor.h
@@ -60,7 +60,7 @@ private:
     bool p_init(AbstractPlanNode*, const ExecutorVector& executorVector);
     bool p_execute(const NValueArray &params);
 
-    static int s_modifiedTuples;
+    static int64_t s_modifiedTuples;
 };
 
 }
